% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_papers}
\alias{analyze_papers}
\title{Analyze papers}
\usage{
analyze_papers(data)
}
\arguments{
\item{data}{dataframe from load_raw_data()}
}
\value{
list with statistics
}
\description{
Analyze papers
}
