% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{get_top_words_by_tag}
\alias{get_top_words_by_tag}
\title{Get top-N words per topic}
\usage{
get_top_words_by_tag(data, n = 5)
}
\arguments{
\item{data}{data.frame with columns \code{abstract} and \code{tag}}

\item{n}{integer, number of words per topic}
}
\value{
tibble with columns \code{tag}, \code{word}, \code{n}
}
\description{
Get top-N words per topic
}
