% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow_client.R
\name{ml_service_is_healthy}
\alias{ml_service_is_healthy}
\title{Check if ML service is healthy}
\usage{
ml_service_is_healthy(ml_service_url = NULL)
}
\arguments{
\item{ml_service_url}{URL of the ML classification service}
}
\value{
logical TRUE if service is responding
}
\description{
Check if ML service is healthy
}
