% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_data.R
\name{search_papers}
\alias{search_papers}
\title{Search papers}
\usage{
search_papers(data, query = NULL, year = NULL)
}
\arguments{
\item{data}{dataframe from load_raw_data()}

\item{query}{text to search}

\item{year}{year filter (optional)}
}
\value{
filtered dataframe
}
\description{
Search papers
}
