% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_client.R
\name{training_cancel_job}
\alias{training_cancel_job}
\title{Cancel a job}
\usage{
training_cancel_job(job_id, ml_service_url = NULL)
}
\arguments{
\item{job_id}{Job identifier returned by one of the \code{training_start_*} calls.}

\item{ml_service_url}{Service URL. Defaults to env var \code{ML_SERVICE_URL}.}
}
\description{
Cancel a job
}
