% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_client.R
\name{training_get_config_safe}
\alias{training_get_config_safe}
\title{Like training_get_config but returns the full call result (ok/data/error).}
\usage{
training_get_config_safe(ml_service_url = NULL)
}
\arguments{
\item{ml_service_url}{Service URL. Defaults to env var \code{ML_SERVICE_URL}.}
}
\description{
Like training_get_config but returns the full call result (ok/data/error).
}
