% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_client.R
\name{training_reload_models}
\alias{training_reload_models}
\title{Reload all .pt models in the inference service.}
\usage{
training_reload_models(ml_service_url = NULL)
}
\arguments{
\item{ml_service_url}{Service URL. Defaults to env var \code{ML_SERVICE_URL}.}
}
\description{
Reload all .pt models in the inference service.
}
