% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_data.R
\name{classify_data}
\alias{classify_data}
\title{Classify data}
\usage{
classify_data(data)
}
\arguments{
\item{data}{dataframe from load_raw_data()}
}
\value{
list with statistics
}
\description{
Classify data
}
