% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_data.R
\name{etl}
\alias{etl}
\title{ETL}
\usage{
etl(cmax_results)
}
\arguments{
\item{cmax_results}{How many papers to load (integer).}
}
\description{
Pipeline that fetches papers from arXiv API (Atom) converts it by classifing and updates database.
}
