% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{get_top_words_by_tag}
\alias{get_top_words_by_tag}
\title{Получить топ-N слов по каждой теме}
\usage{
get_top_words_by_tag(data, n = 5)
}
\arguments{
\item{data}{data.frame с колонками \code{abstract} и \code{tag}}

\item{n}{integer, сколько слов на тему}
}
\value{
tibble с колонками \code{tag}, \code{word}, \code{n}
}
\description{
Получить топ-N слов по каждой теме
}
