% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{launch_app}
\alias{launch_app}
\title{CyberArXiv Shiny GUI}
\usage{
launch_app(
  host = "0.0.0.0",
  port = 3838,
  ml_service_url = NULL,
  launch_browser = interactive()
)
}
\arguments{
\item{host}{Host (default "0.0.0.0").}

\item{port}{Port (default 3838).}

\item{ml_service_url}{URL ML-сервиса.}

\item{launch_browser}{Открывать браузер.}
}
\description{
CyberArXiv Shiny GUI
}
